ajusta banner da loja

+++ src/Divisima.API/Models/CarrinhoCompraItem.cs
@@ -6,9 +6,9 @@ namespace Divisima.Models
     public class CarrinhoCompraItem
     {
         [Key]
-        public int CarrinhoCompraId { get; set; }
+        public long CarrinhoCompraId { get; set; }
         public int Quantidade { get; set; }
-        public int ProdutoId { get; set; }
+        public long ProdutoId { get; set; }
         public string UsuarioId { get; set; }
         public virtual Produto Produto { get; set; }
         public virtual Usuario Usuario { get; set; }
