+++ src/Divisima.API/Models/FotoProduto.cs
@@ -7,10 +7,10 @@ namespace Divisima.Models
     public class FotoProduto
     {
         [Key]
-        public int FotoId { get; set; }
+        public long FotoId { get; set; }
         public string Foto { get; set; }
         public BooleanoEnum IsFotoPrincipal { get; set; }
-        public int ProdutoId { get; set; }
+        public long ProdutoId { get; set; }
         public virtual Produto Produto { get; set; }
     }
 }
