+++ src/Divisima.API/Models/PedidoItem.cs
@@ -7,10 +7,10 @@ namespace Divisima.Models
     public class PedidoItem
     {
         [Key]
-        public int PedidoItemId { get; set; }
+        public long PedidoItemId { get; set; }
         public int Quantidade { get; set; }
-        public int ProdutoId { get; set; }
-        public int PedidoId { get; set; }
+        public long ProdutoId { get; set; }
+        public long PedidoId { get; set; }
         public ProdutoTamanhoEnum Tamanho { get; set; }
         public virtual Produto Produto { get; set; }
         public virtual Pedido Pedido { get; set; }
