+++ src/Divisima.API/Repository/Contracts/IProdutoRepository.cs
@@ -8,12 +8,12 @@ namespace divisima.Repository.Contracts
     {
          Task<List<Produto>> GetAll(int numberPage = 0, int limit = 5);
          Task Cadastrar(Produto produto);
-         Task<Produto> GetById(int id);
+         Task<Produto> GetById(long id);
          Task Editar(Produto produto);
          Task Deletar(Produto produto);
          Task<List<Produto>> GetProductosRecentes(int numberResults = 5);
-         Task<List<Produto>> GetProdutosByCatergoriaId(int categoriaId, int numberPage = 1, int limit = 5);
-         Task BaixarQuantidadeProduto(int produtoId, int quantidade);
+         Task<List<Produto>> GetProdutosByCatergoriaId(long categoriaId, int numberPage = 1, int limit = 5);
+         Task BaixarQuantidadeProduto(long produtoId, int quantidade);
          Task<List<Produto>> BuscarProdutosByName(string nomeProduto);
          Task<Produto> GetLastProdutoCadastrado();
     }
